Creating buttons with onclick attributes
Using onclick attributes is generally not best practice but in this case it meant I could fairly easily reuse some code from my TeaRounder project to create the buttons for now. I can improve that old code at a later date when this project is usable, or if it starts to cause issues.