Tab Button!
I recently started using btnCode to write btnCode, and one of the frustrating things has been that the only way to put tabs in manually is to paste a tab in, because using tab in the browser moves the focus on away from the textarea instead of inserting a tab character. I had been thinking of overriding the keypress, but a really  quick, easy and effective solution is to have a tab button. Of course it would still be good to override the tab keypress but I think having both is what I  would want in the longrun anyway