group locked bug fixed
also btnTypes placed as class in DOM now.

Fixed the bug where groups would lock out if they where open when you changed lang - we simply detect if the data says a new btn to be drawn is an openGroup and flips the data and its type to being a regular group.

I was originally going to make it create the buttons but I got into a bit of a muddle with things, and then realised both ways of fixing this bug are probably about equal from a UX perspective, so long as it is consistent. At a later stage it would be cool to have an option that sets the behaviour the other way, so that it is a preference of the user. For now though that is extra coding with less benefit than other things that I want to add in soon.