Table group and btn
Regarding the revert and commit it reverted: I was going to group a bunch of group btns together whilte adding this table group, thinking that there are now quite a few btns at the top level of HTML, but there were some issues.
a)I am not convinced the reduction in btns onscreen is worth the trade-off of having things buried further
b) closing a group currently only removes the immediate child buttons from the DOM, so that would effectively introduce a bug with nested btns getting stranded if an ancestor group btn is closed while the parent group is still open. I could fix that issue now, but it is very low priority, and the task at hand for today is really the Downloader.

I still wanted the table group and codeBtn though!