iframe links! + boilerplate btn includes charset
charset added because while testing the iframe with some editor test HTML that had emojis, the emojis where showing as other characters. I initially thought it was something wrong with the iframe link but then found it was happening in in new tab links as well but not in downloaded links. Eventually out this together to realise the issue was in the HTML I was testing not the viewer links or extraction. I had used the boiler plate btn to throw something together quickly and that had no charset, and without the charset browsers have to guess - hence different ways of showing the same file gave different results!!