Signalling areas that are a bit undercooked
I like to have things in early just about useable stages shown, but it is a good idea to signal what things are a bit weak at the moment. C# only has the hello world button, and php is not much better right now.  Python and JS only just about make the cut to not have 🚧 symbols on them