Better wording on the styler
I spent ages trying to overcome the pasting rich text problem by approaches involving the element users interact with being a div or a textarea and then clicking apply to move it via a converter but the conversions nothing seemed to work. I will try intercepting the paste event next!