Take 2: Transitions should not effect resizing
.. but they should still happen elsewhere!!

(that last commit, I tried to fix a new issue with resizing the text area, and sort of fixed it, but not how I thought I was fixing it! I was actually wiping over the transition rule on everything, not adding an additional rule. It also turns out that it is the height and width properties changing that need to be instant. - I really did not think those changes through thoroughly! - one of those fixes where you say "oh yeah I remember how CSS actually works" and then promptly facepalm