ghostly style + pickStyle good code
Changes:
new ghostly style: a bit of fun and maybe inspiring for some future users
Replaced that awful if/else chain in function pickStyle() before it got too far out of hand, with some nice pattern matching code that should be easier to keep up to date.

Making of story:
I was doing some more playing around with styles, just having some fun, as you do (its a Sunday evening, and this is a passion project), found a really cool shade of bluey grey #223344 that I felt was cool and deserved a theme based round it, decided it was a ghostly kind of colour so went all in on that theme idea, just in the styler at first (as has become my custom). Then when saving it, was fed up of altering the if/else chain and starting looking into how to derive a var from a pattern. I quickly found something that would work online (w/out the potential security issues of eval) and got it to work, and put in a description comment since I am fairly sure future me will forget some of the detail of how it works while wanting to reuse that tactic for something else someday.