Add btn prototype 1a UX improvement pt1
The UX of the part of the btn adder used for the basic properties of a new btn was frustrating me. I generally derive the jsKey from the text on the btn (and actually I would like that to be filled for me, so I just check it and adjust, rather than copying text all the d*mn time, but that can be done in a future commit). I also want (and now have) the lvlInt choice to be in the actual terms I think of it: do I want it shown immediately or not