Refactoring for Separation of Concerns pt1 data
codebase was getting bloated and awkward to work with. In other words it needed a bit of a refactor. It is mostly the inaccurately named "btnCodeUi.js" that is the big issue. I figure I am going to refactor in stages of commits, so that I can check for any impact in  performance as I go, and roll back anything and readjust anything that has issues.