more tidy up & refactoring: copyText in particular
(in line number order)
Extra comments and spacing in pageLoad to make everything more organised and easy to read at a glance

line 259: a few changes to describe here:
 - further down I refactored function copyText() to function elValueToClipboard(id) which meant that addButtons was not going to work here, I needed drawBtn for the flexibility to specify what to pass to the function in the click event
 - the id of the btn is now btnCopyAllFromEditor rather than an automated conversion of the text on the button
 - I changed the text on the button to specify where it copies from, to make it clear that this button does not copy styles in the styler or anything else
 - Now only 1 call to addButtons exists, making it close to being possibly to scrap that inferior function!!

moved the call of stylerPlainPasteEventOverwrite(); in pageLoad beneath the PostHog stuff, since the PostHog stuff has an impact on perceived page load time whereas stylerPlainPasteEventOverwrite(); does not - also stuck a todo on there for the next commit, as I think there is more I can do there, but this is already quite a chunk of changes to commit

moved function stylerPlainPasteEventOverwrite() up since it is only used in relation to pageLoad

improved the jsdoc on languageChange so that it is extra clear that you pass the actual function and not just the name

lines 514/517: commented out some debug logging code that is not needed currently

lines 576 - 588 refactored and moved copyText() from line 656, into elValueToClipboard(elementID)
 - renamed to make it clear that the desired result of the function is text in the clipboard
 - made it generic, since I would like to have clipboard writing buttons for the styler and add buttons area at some point soon too
 - removed duplicated code (checks for permissions done on pageLoad and not really needed each time as well)
 - added a jsDoc that reminds the reader that 'clipboard-write' permission is required for it to work and that we only draw btns that call this after checking that permission

removed addAllLangBtns() since it was not in use, and would attempt to call a function addLangBtns that is not found in the codebase. Seems to be a stub of a different version of showLangBtns (which seems perfectly fine, and not in any obvious need of replacing).