Ghostly style becomes a little more ghostly
I just wanted to code something fun this evening with near instant gratification. Improving a btncode style never fails to hit that spot!

box shadow on the styler buttons was not visible so I made it a bit more visible (with a slightly red tone), and also added some box shadow to all other buttons that will give a fuzzy ghost trail effect.