intro text hide-able in menu
Intro text sometimes feels like it is in the way, particularly on mobile.

Initially looked at having a little button between the intro text and the rest of the top bar of btns. Then  I realised that if I put the toggle button in the menu, the top bar of btns would be nice and flush against the header area.

Something to consider is to maybe have that intro text in a modal that is open on launch. This has a few potential benefits: 1) could hide other parts of the page being loaded in and styled
2) could let me make the intro text bigger and more nicely styled