blog layout starting point
Colour stripes with a nice green hue to start. Based off saturatedColors.html then moved the colours so that the first hue is a green colour that is a better landing colour than red. Concept is that you have 4 collapsed article areas on screen at start, and then the viewer can expand them to read each one